refactor: centralize test helpers in settings subdirectory
