+++ .trellis/f/F-end-to-end-tests-for-roles/t/open/T-implement-default-roles.md
@@ -39,7 +39,7 @@ import {
   setupRolesTestSuite,
   openRolesSection,
   waitForRolesList,
-} from "./index";
+} from \"../../../helpers\";
 
 test.describe("Feature: Roles Section - Default Roles Loading", () => {
   const testSuite = setupRolesTestSuite();
