+++ tests/desktop/features/settings/llm-setup/index.ts
@@ -1,15 +1,15 @@
 // Types
-export type { Provider } from "./Provider";
-export type { MockLlmConfig } from "./MockLlmConfig";
-export type { StoredLlmConfig } from "./StoredLlmConfig";
+export type { Provider } from "../../../helpers/settings/Provider";
+export type { MockLlmConfig } from "../../../helpers/settings/MockLlmConfig";
+export type { StoredLlmConfig } from "../../../helpers/settings/StoredLlmConfig";
 
 // Helper functions
-export { cleanupLlmStorage } from "./cleanupLlmStorage";
-export { createMockOpenAiConfig } from "./createMockOpenAiConfig";
-export { createMockAnthropicConfig } from "./createMockAnthropicConfig";
-export { waitForEmptyState } from "./waitForEmptyState";
-export { waitForConfigurationList } from "./waitForConfigurationList";
-export { setupLlmTestSuite } from "./setupLlmTestSuite";
+export { cleanupLlmStorage } from "../../../helpers/settings/cleanupLlmStorage";
+export { createMockOpenAiConfig } from "../../../helpers/settings/createMockOpenAiConfig";
+export { createMockAnthropicConfig } from "../../../helpers/settings/createMockAnthropicConfig";
+export { waitForEmptyState } from "../../../helpers/settings/waitForEmptyState";
+export { waitForConfigurationList } from "../../../helpers/settings/waitForConfigurationList";
+export { setupLlmTestSuite } from "../../../helpers/settings/setupLlmTestSuite";
 
 // Re-export from helpers
-export { openLlmSetupSection } from "../../../helpers/openLlmSetupSection";
+export { openLlmSetupSection } from "../../../helpers";
