--- tests/desktop/features/settings/roles/index.ts
@@ -1,13 +1,0 @@
-export { cleanupRolesStorage } from "./cleanupRolesStorage";
-export { openRolesSection } from "./openRolesSection";
-export { setupRolesTestSuite } from "./setupRolesTestSuite";
-export {
-  waitForRolesList,
-  waitForRolesEmptyState,
-  waitForRole,
-} from "./waitForRolesList";
-export {
-  waitForRoleModal,
-  waitForDeleteDialog,
-  waitForModalToClose,
-} from "./waitForRoleModal";
