+++ tests/desktop/helpers/index.ts
@@ -2,7 +2,31 @@
  * Barrel file for test helpers
  */
 export { createElectronApp } from "./createElectronApp";
-export { openAppearanceSettings } from "./openAppearanceSettings";
+export { openAdvancedSettings } from "./settings/openAdvancedSettings";
+export { openAppearanceSettings } from "./settings/openAppearanceSettings";
+export { openLlmSetupSection } from "./settings/openLlmSetupSection";
+export { cleanupLlmStorage } from "./settings/cleanupLlmStorage";
+export { cleanupRolesStorage } from "./settings/cleanupRolesStorage";
+export { createMockAnthropicConfig } from "./settings/createMockAnthropicConfig";
+export { createMockOpenAiConfig } from "./settings/createMockOpenAiConfig";
+export { setupLlmTestSuite } from "./settings/setupLlmTestSuite";
+export { waitForConfigurationList } from "./settings/waitForConfigurationList";
+export { waitForEmptyState } from "./settings/waitForEmptyState";
+export { openRolesSection } from "./settings/openRolesSection";
+export { setupRolesTestSuite } from "./settings/setupRolesTestSuite";
+export {
+  waitForRolesList,
+  waitForRolesEmptyState,
+  waitForRole,
+} from "./settings/waitForRolesList";
+export {
+  waitForRoleModal,
+  waitForDeleteDialog,
+  waitForModalToClose,
+} from "./settings/waitForRoleModal";
 export type { TestElectronApplication } from "./TestElectronApplication";
 export type { TestHelpers } from "./TestHelpers";
 export type { TestWindow } from "./TestWindow";
+export type { MockLlmConfig } from "./settings/MockLlmConfig";
+export type { Provider } from "./settings/Provider";
+export type { StoredLlmConfig } from "./settings/StoredLlmConfig";
