+++ tests/desktop/helpers/settings/MockLlmConfig.ts
