+++ tests/desktop/helpers/settings/Provider.ts
