+++ tests/desktop/helpers/settings/StoredLlmConfig.ts
