+++ tests/desktop/helpers/settings/cleanupRolesStorage.ts
