+++ tests/desktop/helpers/settings/createMockAnthropicConfig.ts
