+++ tests/desktop/helpers/settings/createMockOpenAiConfig.ts
