+++ tests/desktop/helpers/settings/openAdvancedSettings.ts
