+++ tests/desktop/helpers/settings/openAppearanceSettings.ts
