+++ tests/desktop/helpers/settings/openLlmSetupSection.ts
