+++ tests/desktop/helpers/settings/setupLlmTestSuite.ts
@@ -5,7 +5,7 @@ import {
   createElectronApp,
   type TestElectronApplication,
   type TestWindow,
-} from "../../../helpers";
+} from "../index";
 import { cleanupLlmStorage } from "./cleanupLlmStorage";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -21,7 +21,7 @@ export const setupLlmTestSuite = () => {
     // Launch Electron app with test environment
     const electronPath = path.join(
       __dirname,
-      "../../../../../apps/desktop/dist-electron/electron/main.js",
+      "../../../../apps/desktop/dist-electron/electron/main.js",
     );
     electronApp = await createElectronApp(electronPath);
