+++ tests/desktop/helpers/settings/setupRolesTestSuite.ts
@@ -5,7 +5,7 @@ import {
   createElectronApp,
   type TestElectronApplication,
   type TestWindow,
-} from "../../../helpers";
+} from "../index";
 import { cleanupRolesStorage } from "./cleanupRolesStorage";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -20,7 +20,7 @@ export const setupRolesTestSuite = () => {
     // Launch Electron app with test environment
     const electronPath = path.join(
       __dirname,
-      "../../../../../apps/desktop/dist-electron/electron/main.js",
+      "../../../../apps/desktop/dist-electron/electron/main.js",
     );
     electronApp = await createElectronApp(electronPath);
