+++ tests/desktop/helpers/settings/waitForConfigurationList.ts
@@ -1,5 +1,5 @@
 import { expect } from "@playwright/test";
-import type { TestWindow } from "../../../helpers";
+import type { TestWindow } from "../index";
 
 export const waitForConfigurationList = async (window: TestWindow) => {
   await expect(window.locator('[role="article"]').first()).toBeVisible({
